City population topojson progess
Have the correct SQL and nodeJS to call it.
We now set a boolean next to every city that says whether it's a landing page city.
If it is, add it to the landing page topojson.
The topojson function we run is mostly done.
Just need to get more city information from GEOJson to topojson.